Assume that libcheri directories always exist
Currently there are already quite a few places that hardcode libcheri
(since they assumed a hybrid rootfs/sysroot as that was the only thing
that existed at the time).

https://github.com/CTSRD-CHERI/cheribsd/pull/548 adds a libcheri symlink
for the native ABI == purecap case. This will unbreak those uses of
libcheri and allows for some minor simplifications in cheribuild test scripts.